+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `eways-client` will be documented in this file
 
+## 1.1.0 - 2019-02-27
+
+- Add getProduct function
+- Add setProduct function
+- Add getRequestResponse function
+- Add setRequestResponse function
+
 ## 1.0.0 - 2019-02-25
 
 - initial release